Resolving "TypeError: integer argument expected, got float" by adding int to become (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))

+++ pygameMultibot2.py
@@ -30,7 +30,7 @@ class Robot():
         self.prev2Load=0
         self.image = pygame.image.load(r"Resources/Roomba-bot.png")
 #        width, height = self.image.get_width(), self.image.get_height()  # get size of image
-        self.image = pygame.transform.scale(self.image, (GRID_SIZE*SCALE, GRID_SIZE*SCALE))
+        self.image = pygame.transform.scale(self.image, (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))
         self.id = 1 + rbtCount*2
         warehouseFloor[self.x][self.y][1] = 1 + rbtCount*2 + self.loaded
         self.interval = 0
@@ -87,7 +87,7 @@ class Parcel():
         #   self.y = random.randint(0, ROW_COUNT-1)
         self.image = pygame.image.load(r"Resources/package.png")
 #        width, height = self.image.get_width(), self.image.get_height()  # get size of image
-        self.image = pygame.transform.scale(self.image, (GRID_SIZE*SCALE, GRID_SIZE*SCALE))
+        self.image = pygame.transform.scale(self.image, (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))
         warehouseFloor[self.x][self.y][0] = 1
 
 
@@ -102,7 +102,7 @@ class Destination():
         self.y = y
         self.image = pygame.image.load(r"Resources/desti.png")
 #        width, height = self.image.get_width(), self.image.get_height()  # get size of image
-        self.image = pygame.transform.scale(self.image, (GRID_SIZE*SCALE, GRID_SIZE*SCALE))
+        self.image = pygame.transform.scale(self.image, (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))
         warehouseFloor[self.x][self.y][0] = 2
 
 
@@ -118,7 +118,7 @@ class Boulder():
         self.y = y
         self.image = pygame.image.load(r"Resources/brick.png")
 #        width, height = self.image.get_width(), self.image.get_height()  # get size of image
-        self.image = pygame.transform.scale(self.image, (GRID_SIZE*SCALE, GRID_SIZE*SCALE))
+        self.image = pygame.transform.scale(self.image, (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))
         warehouseFloor[self.x][self.y][0] = 0.5
 
 
@@ -129,7 +129,7 @@ class Human():
         self.direction = 1 #1 is right side -1 is left
         self.image = pygame.image.load(r"Resources/person.png")
 #        width, height = self.image.get_width(), self.image.get_height()  # get size of image
-        self.image = pygame.transform.scale(self.image, (GRID_SIZE*SCALE, GRID_SIZE*SCALE))
+        self.image = pygame.transform.scale(self.image, (int(GRID_SIZE*SCALE), int(GRID_SIZE*SCALE)))
         warehouseFloor[self.x][self.y][0] = 1.5
 
     def randMov(self, warehouseFloor):
